# Optimize IRenderable system to reduce unnecessary computation