--- a/src/cgame/etj_crosshair.h
+++ b/src/cgame/etj_crosshair.h
@@ -66,6 +66,6 @@ protected:
 public:
   Crosshair();
   void render() const override;
-  void beforeRender() override;
+  bool beforeRender() override;
 };
 } // namespace ETJump