--- a/src/cgame/etj_irenderable.h
+++ b/src/cgame/etj_irenderable.h
@@ -27,7 +27,7 @@ namespace ETJump {
 class IRenderable {
 public:
   virtual ~IRenderable(){};
-  virtual void beforeRender() = 0;
+  virtual bool beforeRender() = 0;
   virtual void render() const = 0;
 };
 } // namespace ETJump