--- a/src/cgame/etj_snaphud.cpp
+++ b/src/cgame/etj_snaphud.cpp
@@ -222,7 +222,11 @@ void Snaphud::UpdateMaxSnapZones(float wishspeed, pmove_t *pm) {
   snap.absAccel.resize(maxSnaphudZonesQ1);
 }
 
-void Snaphud::beforeRender() {
+bool Snaphud::beforeRender() {
+  if (canSkipDraw()) {
+    return false;
+  }
+
   const int8_t uCmdScale =
       ps->stats[STAT_USERCMD_BUTTONS] & (BUTTON_WALKING << 8)
           ? CMDSCALE_WALK
@@ -231,6 +235,13 @@ void Snaphud::beforeRender() {
 
   // get correct pmove state
   pm = PmoveUtils::getPmove(cmd);
+
+  // water and ladder movement are not important
+  // since speed is capped anyway
+  // check this only after we have a valid pmove
+  if (pm->pmext->waterlevel > 1 || pm->pmext->ladder) {
+    return false;
+  }
 
   // show upmove influence?
   float scale = etj_snapHUDTrueness.integer &
@@ -273,12 +284,11 @@ void Snaphud::beforeRender() {
 
   edgesOnly = etj_drawSnapHUD.integer == 2;
   edgeThickness = Numeric::clamp(etj_snapHUDEdgeThickness.integer, 1, 128);
+
+  return true;
 }
 
 void Snaphud::render() const {
-  if (canSkipDraw()) {
-    return;
-  }
 
   float h = etj_snapHUDHeight.value;
   float y = 240 + etj_snapHUDOffsetY.value;
@@ -455,12 +465,6 @@ bool Snaphud::canSkipDraw() const {
 
   if (BG_PlayerMounted(ps->eFlags) || ps->weapon == WP_MOBILE_MG42_SET ||
       ps->weapon == WP_MORTAR_SET) {
-    return true;
-  }
-
-  // water and ladder movement are not important
-  // since speed is capped anyway
-  if (pm->pmext->waterlevel > 1 || pm->pmext->ladder) {
     return true;
   }