--- a/src/cgame/etj_speed_drawable.h
+++ b/src/cgame/etj_speed_drawable.h
@@ -52,6 +52,6 @@ public:
   DisplaySpeed();
   ~DisplaySpeed();
   void render() const override;
-  void beforeRender() override;
+  bool beforeRender() override;
 };
 } // namespace ETJump