`IRenderable::beforeRender` is now a boolean function, and `IRenderable::render` will not execute if it returns false. This makes it easier to optimize drawing functions by reducing the amount of complex math that needs to be done to setup renderable drawing in beforeRender, since we can just easily return false now to skip most of the stuff that's being done, and just skip the render that way.

Tested with `timedemo`, a random demo I had from a full map timerun of sunrise with toggling the following cvars (the most complex ones to render), rough average of five runs:

* `etj_drawCGaz`
* `etj_drawSnapHUD`
* `etj_drawJumpSpeeds`
* `etj_drawStrafeQuality`
* `etj_drawUpmoveMeter`

FPS before refactor:
* renderables off: ~1080fps
* renderables on: ~1050fps

FPS after refactor:
* renderables off: ~1350fps
* rennderables on: ~1060fps

This also fixes ob detector/watcher drawing while scoreboard is up, a small bug as a result of the recent refactor